pi_dashcam/requirements_pi.txt
smbus2
RPi.GPIO
Pillow
